pam: realm_file can contain tokens
The realm_file parameter now understands the special tokens
'~' (user home directory) and '%u' (username). '~~' can be used
for a single tilde character. '%%' can be used for a single
percent character. Unknown tokens (such as '%a') are copied
verbatim.

The realm file is no longer assumed to be in the users home
directory. Use '~/.otp_token' instead.