Make authentication with challenge work.
When handling a server challenge, pam_linotp_validate_password()
will call pam_linotp_get_authtok() with "token_length" set to 0
(unconditionally).  pam_linotp_get_authtok() passes this on to
pam_linotp_get_authtok_no_use_first_pass(), which will just
error out in that case, doing nothing.

Change pam_linotp_get_authtok_no_use_first_pass() to always do its
job, and only check token_length when deciding whether or not to
write back the token length to the caller (which is the only way
this is used in the other caller, pam_sm_authenticate()).

The whole purpose of the integer use of "token_len" in this
context remains unclear (as the value written back is never used), but
passing in always-0 and then always erroring-out is never useful.

Signed-off-by: Gert Doering <gert@space.net>